WIK-1156 Replace failing org.apache.commons:commons-lang3
The 3.7 version of commons-lang3 has Java version enum constants only up to Java 10.
When faced with Java 11 it fails badly with a NullPointerException.
By replacing this dependency with a newer version that understands
the new Java naming conventions, we can avoid the NPE.
Also added properties for hard coded aws-java-sdk and jackson versions.
Updated cqlVersion for cassandra-maven-plugin for dependency on cassandra-all.